+++ modules/service/src/test/scala/lucuma/odb/graphql/query/ExecutionTestSupport.scala
@@ -4,47 +4,16 @@
 package lucuma.odb.graphql
 package query
 
-import cats.data.NonEmptySet
 import cats.effect.Clock
 import cats.effect.IO
 import cats.syntax.all.*
-import eu.timepit.refined.types.numeric.PosInt
-import eu.timepit.refined.types.numeric.PosLong
 import io.circe.Json
 import io.circe.literal.*
-import io.circe.syntax.*
-import lucuma.core.enums.Breakpoint
 import lucuma.core.enums.CalibrationRole
 import lucuma.core.enums.DatasetQaState
 import lucuma.core.enums.DatasetStage
-import lucuma.core.enums.Flamingos2Decker
-import lucuma.core.enums.Flamingos2Disperser
-import lucuma.core.enums.Flamingos2Filter
-import lucuma.core.enums.Flamingos2Fpu
-import lucuma.core.enums.Flamingos2LyotWheel
-import lucuma.core.enums.Flamingos2ReadMode
-import lucuma.core.enums.Flamingos2ReadoutMode
-import lucuma.core.enums.Flamingos2Reads
-import lucuma.core.enums.GcalArc
-import lucuma.core.enums.GcalBaselineType
-import lucuma.core.enums.GcalContinuum
-import lucuma.core.enums.GcalDiffuser
-import lucuma.core.enums.GcalFilter
-import lucuma.core.enums.GcalShutter
-import lucuma.core.enums.GmosAmpCount
-import lucuma.core.enums.GmosAmpGain
-import lucuma.core.enums.GmosAmpReadMode
-import lucuma.core.enums.GmosDtax
-import lucuma.core.enums.GmosGratingOrder
-import lucuma.core.enums.GmosNorthFilter
-import lucuma.core.enums.GmosNorthFpu
-import lucuma.core.enums.GmosNorthGrating
-import lucuma.core.enums.GmosRoi
-import lucuma.core.enums.GmosXBinning
-import lucuma.core.enums.GmosYBinning
 import lucuma.core.enums.StepGuideState
 import lucuma.core.math.Angle
-import lucuma.core.math.BoundedInterval
 import lucuma.core.math.Offset
 import lucuma.core.math.Wavelength
 import lucuma.core.math.WavelengthDither
@@ -55,17 +24,8 @@ import lucuma.core.model.User
 import lucuma.core.model.sequence.Dataset
 import lucuma.core.model.sequence.InstrumentExecutionConfig
 import lucuma.core.model.sequence.Step
-import lucuma.core.model.sequence.StepConfig
-import lucuma.core.model.sequence.StepConfig.Gcal
 import lucuma.core.model.sequence.TelescopeConfig
-import lucuma.core.model.sequence.flamingos2.Flamingos2DynamicConfig
-import lucuma.core.model.sequence.flamingos2.Flamingos2FpuMask
-import lucuma.core.model.sequence.gmos.DynamicConfig.GmosNorth
-import lucuma.core.model.sequence.gmos.GmosCcdMode
-import lucuma.core.model.sequence.gmos.GmosFpuMask
-import lucuma.core.model.sequence.gmos.GmosGratingConfig
 import lucuma.core.syntax.string.*
-import lucuma.core.syntax.timespan.*
 import lucuma.core.util.TimeSpan
 import lucuma.core.util.Timestamp
 import lucuma.odb.data.OdbError
@@ -74,12 +34,7 @@ import lucuma.odb.logic.Generator
 import lucuma.odb.logic.TimeEstimateCalculatorImplementation
 import lucuma.odb.sequence.util.CommitHash
 import lucuma.odb.service.Services
-import lucuma.odb.smartgcal.data.Flamingos2
-import lucuma.odb.smartgcal.data.Gmos
-import lucuma.odb.smartgcal.data.SmartGcalValue
-import lucuma.odb.smartgcal.data.SmartGcalValue.LegacyInstrumentConfig
 import natchez.Trace.Implicits.noop
-import skunk.Session
 
 trait ExecutionTestSupport extends OdbSuite with ObservingModeSetupOperations {
 
@@ -96,145 +51,6 @@ trait ExecutionTestSupport extends OdbSuite with ObservingModeSetupOperations {
 
   val createProgram: IO[Program.Id] =
     createProgramAs(pi, "Sequence Testing")
-
-  val f2_key_JH1: Flamingos2.TableKey =
-    Flamingos2.TableKey(
-      Flamingos2Disperser.R1200JH.some,
-      Flamingos2Filter.JH,
-      Flamingos2Fpu.LongSlit1.some
-    )
-
-  val f2_flat_JH1 =
-    SmartGcalValue(
-      Gcal(
-        Gcal.Lamp.fromContinuum(GcalContinuum.IrGreyBodyHigh),
-        GcalFilter.Nd20,
-        GcalDiffuser.Ir,
-        GcalShutter.Open
-      ),
-      GcalBaselineType.Night,
-      PosInt.unsafeFrom(1),
-      LegacyInstrumentConfig(
-        TimeSpan.unsafeFromMicroseconds(15_000_000L)
-      )
-    )
-
-  val f2_arc_JH1 =
-    SmartGcalValue(
-      Gcal(
-        Gcal.Lamp.fromArcs(NonEmptySet.one(GcalArc.ArArc)),
-        GcalFilter.Nir,
-        GcalDiffuser.Ir,
-        GcalShutter.Closed
-      ),
-      GcalBaselineType.Night,
-      PosInt.unsafeFrom(1),
-      LegacyInstrumentConfig(
-        TimeSpan.unsafeFromMicroseconds(32_000_000L)
-      )
-    )
-
-  val gn_key_0_50: Gmos.TableKey[GmosNorthGrating, GmosNorthFilter, GmosNorthFpu] =
-    Gmos.TableKey(
-      Gmos.GratingConfigKey(
-        GmosNorthGrating.R831_G5302,
-        GmosGratingOrder.One,
-        BoundedInterval.unsafeOpenUpper(Wavelength.Min, Wavelength.Max)
-      ).some,
-      GmosNorthFilter.RPrime.some,
-      GmosNorthFpu.LongSlit_0_50.some,
-      GmosXBinning.One,
-      GmosYBinning.Two,
-      GmosAmpGain.Low
-    )
-
-  // NB: 2x4, no filter
-  val gn_key_1_00: Gmos.TableKey[GmosNorthGrating, GmosNorthFilter, GmosNorthFpu] =
-    Gmos.TableKey(
-      Gmos.GratingConfigKey(
-        GmosNorthGrating.R831_G5302,
-        GmosGratingOrder.One,
-        BoundedInterval.unsafeOpenUpper(Wavelength.Min, Wavelength.Max)
-      ).some,
-      none,
-      GmosNorthFpu.LongSlit_1_00.some,
-      GmosXBinning.Two,
-      GmosYBinning.Four,
-      GmosAmpGain.Low
-    )
-
-  val gn_key_5_00: Gmos.TableKey[GmosNorthGrating, GmosNorthFilter, GmosNorthFpu] =
-    Gmos.TableKey(
-      Gmos.GratingConfigKey(
-        GmosNorthGrating.R831_G5302,
-        GmosGratingOrder.One,
-        BoundedInterval.unsafeOpenUpper(Wavelength.Min, Wavelength.Max)
-      ).some,
-      GmosNorthFilter.RPrime.some,
-      GmosNorthFpu.LongSlit_5_00.some,
-      GmosXBinning.One,
-      GmosYBinning.Two,
-      GmosAmpGain.Low
-    )
-
-  val gn_flat =
-    SmartGcalValue(
-      Gcal(
-        Gcal.Lamp.fromContinuum(GcalContinuum.QuartzHalogen5W),
-        GcalFilter.Gmos,
-        GcalDiffuser.Ir,
-        GcalShutter.Open
-      ),
-      GcalBaselineType.Night,
-      PosInt.unsafeFrom(1),
-      LegacyInstrumentConfig(
-        TimeSpan.unsafeFromMicroseconds(1_000_000L)
-      )
-    )
-
-  val gn_arc =
-    SmartGcalValue(
-      Gcal(
-        Gcal.Lamp.fromArcs(NonEmptySet.one(GcalArc.CuArArc)),
-        GcalFilter.None,
-        GcalDiffuser.Visible,
-        GcalShutter.Closed
-      ),
-      GcalBaselineType.Day,
-      PosInt.unsafeFrom(1),
-      LegacyInstrumentConfig(
-        TimeSpan.unsafeFromMicroseconds(1_000_000L)
-      )
-    )
-
-  override def dbInitialization: Option[Session[IO] => IO[Unit]] = Some { s =>
-    val flamingos2TableRows: List[Flamingos2.TableRow] =
-      List(
-        Flamingos2.TableRow(PosLong.unsafeFrom(1), f2_key_JH1, f2_flat_JH1),
-        Flamingos2.TableRow(PosLong.unsafeFrom(1), f2_key_JH1, f2_arc_JH1)
-      )
-
-    val gmosNorthTableRows: List[Gmos.TableRow.North] =
-      List(
-        Gmos.TableRow(PosLong.unsafeFrom(1), gn_key_0_50, gn_flat),
-        Gmos.TableRow(PosLong.unsafeFrom(1), gn_key_0_50, gn_arc),
-        Gmos.TableRow(PosLong.unsafeFrom(1), gn_key_1_00, gn_flat),
-        Gmos.TableRow(PosLong.unsafeFrom(1), gn_key_1_00, gn_arc),
-        Gmos.TableRow(PosLong.unsafeFrom(1), gn_key_5_00, gn_flat),
-        Gmos.TableRow(PosLong.unsafeFrom(1), gn_key_5_00, gn_arc)
-      )
-
-    Enums.load(s).flatMap: e =>
-      val services = Services.forUser(pi /* doesn't matter*/, e, None)(s)
-      services.transactionally:
-        (
-          flamingos2TableRows.zipWithIndex.traverse_ : (r, i) =>
-            services.smartGcalService.insertFlamingos2(i, r)
-        ) *> (
-          gmosNorthTableRows.zipWithIndex.traverse_ : (r, i) =>
-            services.smartGcalService.insertGmosNorth(i, r)
-        )
-  }
 
   def calibrationProgram(role: CalibrationRole): IO[Program.Id] =
     query(
@@ -296,80 +112,6 @@ trait ExecutionTestSupport extends OdbSuite with ObservingModeSetupOperations {
   def twilightTargets: IO[List[Target.Id]] =
     calibrationTargets(CalibrationRole.Twilight)
 
-  val Flamingos2AtomQuery: String =
-    s"""
-      description
-      observeClass
-      steps {
-        instrumentConfig {
-          exposure { seconds }
-          disperser
-          filter
-          readMode
-          lyotWheel
-          fpu { builtin }
-          decker
-          readoutMode
-          reads
-        }
-        stepConfig {
-          stepType
-          ... on Gcal {
-            continuum
-            arcs
-          }
-        }
-        telescopeConfig {
-          offset {
-            p { arcseconds }
-            q { arcseconds }
-          }
-          guiding
-        }
-        observeClass
-        breakpoint
-      }
-    """
-
-  val GmosAtomQuery: String =
-    s"""
-      description
-      observeClass
-      steps {
-        instrumentConfig {
-          exposure { seconds }
-          readout {
-            xBin
-            yBin
-          }
-          roi
-          gratingConfig {
-            grating
-            wavelength { nanometers }
-          }
-          filter
-          fpu { builtin }
-          centralWavelength { nanometers }
-        }
-        stepConfig {
-          stepType
-          ... on Gcal {
-            continuum
-            arcs
-          }
-        }
-        telescopeConfig {
-          offset {
-            p { arcseconds }
-            q { arcseconds }
-          }
-          guiding
-        }
-        observeClass
-        breakpoint
-      }
-    """
-
   def excutionConfigQuery(inst: String, sequenceType: String, atomQuery: String, futureLimit: Option[Int]): String =
     s"""
       execution {
@@ -389,18 +131,6 @@ trait ExecutionTestSupport extends OdbSuite with ObservingModeSetupOperations {
       }
     """
 
-  def flamingos2AcquisitionQuery(futureLimit: Option[Int]): String =
-    excutionConfigQuery("flamingos2", "acquisition", Flamingos2AtomQuery, futureLimit)
-
-  def flamingos2ScienceQuery(futureLimit: Option[Int]): String =
-    excutionConfigQuery("flamingos2", "science", Flamingos2AtomQuery, futureLimit)
-
-  def gmosNorthAcquisitionQuery(futureLimit: Option[Int]): String =
-    excutionConfigQuery("gmosNorth", "acquisition", GmosAtomQuery, futureLimit)
-
-  def gmosNorthScienceQuery(futureLimit: Option[Int]): String =
-    excutionConfigQuery("gmosNorth", "science", GmosAtomQuery, futureLimit)
-
   val ObsWavelength: Wavelength =
     Wavelength.decimalNanometers.unsafeGet(BigDecimal("500.0"))
 
@@ -408,100 +138,6 @@ trait ExecutionTestSupport extends OdbSuite with ObservingModeSetupOperations {
     ObsWavelength.unsafeOffset(
       WavelengthDither.decimalNanometers.unsafeGet(BigDecimal(ditherNm))
     )
-
-  val Flamingos2Science: Flamingos2DynamicConfig =
-    Flamingos2DynamicConfig(
-      fakeItcSpectroscopyResult.exposureTime,
-      Flamingos2Disperser.R1200JH.some,
-      Flamingos2Filter.JH,
-      Flamingos2ReadMode.Bright,
-      Flamingos2LyotWheel.F16,
-      Flamingos2FpuMask.Builtin(Flamingos2Fpu.LongSlit1),
-      Flamingos2Decker.LongSlit,
-      Flamingos2ReadoutMode.Science,
-      Flamingos2ReadMode.Bright.readCount
-    )
-
-  def gmosNorthScience(ditherNm: Int): GmosNorth =
-    GmosNorth(
-      fakeItcSpectroscopyResult.exposureTime,
-      GmosCcdMode(GmosXBinning.One, GmosYBinning.Two, GmosAmpCount.Twelve, GmosAmpGain.Low, GmosAmpReadMode.Slow),
-      GmosDtax.Zero,
-      GmosRoi.FullFrame,
-      GmosGratingConfig.North(GmosNorthGrating.R831_G5302, GmosGratingOrder.One, obsWavelengthAt(ditherNm)).some,
-      GmosNorthFilter.RPrime.some,
-      GmosFpuMask.Builtin(GmosNorthFpu.LongSlit_0_50).some
-    )
-
-  def gmosNorthArc(ditherNm: Int): GmosNorth =
-    gmosNorthScience(ditherNm).copy(exposure = gn_arc.instrumentConfig.exposureTime)
-
-  def gmosNorthFlat(ditherNm: Int): GmosNorth =
-    gmosNorthScience(ditherNm).copy(exposure = gn_flat.instrumentConfig.exposureTime)
-
-  val Flamingos2Acq0: Flamingos2DynamicConfig =
-    Flamingos2Science.copy(
-      exposure  = fakeItcImagingResult.exposureTime,
-      disperser = none,
-      filter    = Flamingos2Filter.H,
-      fpu       = Flamingos2FpuMask.Imaging,
-      decker    = Flamingos2Decker.Imaging
-    )
-
-  val Flamingos2Acq1: Flamingos2DynamicConfig =
-    Flamingos2Acq0.copy(
-      exposure  = 10.secondTimeSpan,
-      fpu       = Flamingos2FpuMask.Builtin(Flamingos2Fpu.LongSlit1),
-      decker    = Flamingos2Decker.LongSlit
-    )
-
-  val Flamingos2Acq2: Flamingos2DynamicConfig =
-    Flamingos2Acq1.copy(
-      exposure = fakeItcImagingResult.exposureTime
-    )
-
-  def flamingos2Acq(step: Int): Flamingos2DynamicConfig =
-    step match
-      case 0 => Flamingos2Acq0
-      case 1 => Flamingos2Acq1
-      case 2 => Flamingos2Acq2
-      case _ => sys.error("Only 3 steps in a Flamingos 2 Acq")
-
-  val GmosNorthAcq0: GmosNorth =
-    gmosNorthScience(0).copy(
-      exposure      = fakeItcImagingResult.exposureTime,
-      readout       = GmosCcdMode(GmosXBinning.Two, GmosYBinning.Two, GmosAmpCount.Twelve, GmosAmpGain.Low, GmosAmpReadMode.Fast),
-      roi           = GmosRoi.Ccd2,
-      gratingConfig = none,
-      filter        = GmosNorthFilter.GPrime.some,
-      fpu           = none
-    )
-
-  val GmosNorthAcq1: GmosNorth =
-    GmosNorthAcq0.copy(
-      exposure = GmosNorthAcq0.exposure *| 2,
-      readout  = GmosCcdMode(GmosXBinning.One, GmosYBinning.One, GmosAmpCount.Twelve, GmosAmpGain.Low, GmosAmpReadMode.Fast),
-      roi      = GmosRoi.CentralStamp,
-      fpu      = gmosNorthScience(0).fpu
-    )
-
-  val GmosNorthAcq2: GmosNorth =
-    GmosNorthAcq1.copy(
-      exposure = GmosNorthAcq0.exposure *| 3
-    )
-
-  def gmosNorthAcq(step: Int): GmosNorth =
-    step match
-      case 0 => GmosNorthAcq0
-      case 1 => GmosNorthAcq1
-      case 2 => GmosNorthAcq2
-      case _ => sys.error("Only 3 steps in a GMOS North Acq")
-
-  val FlatStep: StepConfig.Gcal =
-    StepConfig.Gcal(Gcal.Lamp.fromContinuum(GcalContinuum.QuartzHalogen5W), GcalFilter.Gmos, GcalDiffuser.Ir, GcalShutter.Open)
-
-  val ArcStep: StepConfig.Gcal  =
-    StepConfig.Gcal(Gcal.Lamp.fromArcs(NonEmptySet.one(GcalArc.CuArArc)), GcalFilter.None, GcalDiffuser.Visible, GcalShutter.Closed)
 
   def telescopeConfig(p: Int, q: Int, g: StepGuideState): TelescopeConfig =
     TelescopeConfig(
@@ -525,133 +161,6 @@ trait ExecutionTestSupport extends OdbSuite with ObservingModeSetupOperations {
         "guiding": ${t.guiding}
       }
     """
-
-  protected def flamingos2ExpectedInstrumentConfig(f2: Flamingos2DynamicConfig): Json =
-    json"""
-      {
-        "exposure": { "seconds": ${f2.exposure.toSeconds} },
-        "disperser": ${f2.disperser.fold(Json.Null)(_.asJson)},
-        "filter": ${f2.filter},
-        "readMode": ${f2.readMode},
-        "lyotWheel": ${f2.lyotWheel},
-        "fpu": ${
-          f2.fpu match
-            case Flamingos2FpuMask.Imaging      => Json.Null
-            case Flamingos2FpuMask.Builtin(f)   => json"""{ "builtin": $f }"""
-            case Flamingos2FpuMask.Custom(f, w) => json"""{ "filename": ${f.value}, "slitWidth": $w }"""
-        },
-        "decker": ${f2.decker},
-        "readoutMode": ${f2.readoutMode},
-        "reads": ${f2.reads}
-      }
-    """
-
-  protected def flamingos2ExpectedAcq(step: Int, p: Int, breakpoint: Breakpoint = Breakpoint.Disabled): Json =
-    json"""
-      {
-        "instrumentConfig" : ${flamingos2ExpectedInstrumentConfig(flamingos2Acq(step))},
-        "stepConfig" : { "stepType":  "SCIENCE" },
-        "telescopeConfig": ${expectedTelescopeConfig(p, 0, StepGuideState.Enabled)},
-        "observeClass" : "ACQUISITION",
-        "breakpoint": ${breakpoint.tag.toScreamingSnakeCase.asJson}
-      }
-    """
-
-  protected def gmosNorthExpectedInstrumentConfig(gn: GmosNorth): Json =
-    json"""
-      {
-        "exposure": { "seconds": ${gn.exposure.toSeconds} },
-        "readout": {
-          "xBin": ${gn.readout.xBin},
-          "yBin": ${gn.readout.yBin}
-        },
-        "roi": ${gn.roi},
-        "gratingConfig": ${gn.gratingConfig.fold(Json.Null) { gc =>
-          json"""
-            {
-              "grating": ${gc.grating},
-              "wavelength": {
-                "nanometers": ${gc.wavelength.toNanometers.value.value}
-              }
-            }
-          """
-        }},
-        "filter": ${gn.filter},
-        "fpu": ${gn.fpu.fold(Json.Null) { fpu =>
-          json"""{ "builtin": ${fpu.builtin.map(_.value)} }"""
-        }},
-        "centralWavelength": ${gn.centralWavelength.fold(Json.Null) { cw =>
-          json"""
-            {
-              "nanometers": ${cw.toNanometers.value.value}
-            }
-          """
-        }}
-      }
-    """
-
-  protected def gmosNorthExpectedArc(ditherNm: Int, p: Int, q: Int): Json =
-    json"""
-      {
-        "instrumentConfig" : ${gmosNorthExpectedInstrumentConfig(gmosNorthArc(ditherNm))},
-        "stepConfig" : {
-          "stepType": "GCAL",
-          "continuum" : null,
-          "arcs" : ${gn_arc.gcalConfig.lamp.arcs.map(_.toList) }
-        },
-        "telescopeConfig": ${expectedTelescopeConfig(p, q, StepGuideState.Disabled)},
-        "observeClass" : "NIGHT_CAL",
-        "breakpoint": "DISABLED"
-      }
-    """
-
-  protected def gmosNorthExpectedFlat(ditherNm: Int, p: Int, q: Int): Json =
-    json"""
-      {
-        "instrumentConfig" : ${gmosNorthExpectedInstrumentConfig(gmosNorthFlat(ditherNm))},
-        "stepConfig" : {
-          "stepType": "GCAL",
-          "continuum" : ${gn_flat.gcalConfig.lamp.continuum},
-          "arcs" : []
-        },
-        "telescopeConfig": ${expectedTelescopeConfig(p, q, StepGuideState.Disabled)},
-        "observeClass" : "NIGHT_CAL",
-        "breakpoint": "DISABLED"
-      }
-    """
-
-  protected def gmosNorthExpectedScience(ditherNm: Int, p: Int, q: Int): Json =
-    json"""
-      {
-        "instrumentConfig" : ${gmosNorthExpectedInstrumentConfig(gmosNorthScience(ditherNm))},
-        "stepConfig" : { "stepType": "SCIENCE" },
-        "telescopeConfig": ${expectedTelescopeConfig(p, q, StepGuideState.Enabled)},
-        "observeClass" : "SCIENCE",
-        "breakpoint": "DISABLED"
-      }
-    """
-
-  protected def gmosNorthExpectedAcq(step: Int, p: Int, breakpoint: Breakpoint = Breakpoint.Disabled): Json =
-    json"""
-      {
-        "instrumentConfig" : ${gmosNorthExpectedInstrumentConfig(gmosNorthAcq(step))},
-        "stepConfig" : { "stepType":  "SCIENCE" },
-        "telescopeConfig": ${expectedTelescopeConfig(p, 0, StepGuideState.Enabled)},
-        "observeClass" : "ACQUISITION",
-        "breakpoint": ${breakpoint.tag.toScreamingSnakeCase.asJson}
-      }
-    """
-
-  protected def gmosNorthExpectedScienceAtom(ditherNm: Int, p: Int, q: Int, exposures: Int): Json =
-    val steps = List(
-      gmosNorthExpectedArc(ditherNm, p, q), gmosNorthExpectedFlat(ditherNm, p, q)
-    ) ++ List.fill(exposures)(gmosNorthExpectedScience(ditherNm, p, q))
-
-    Json.obj(
-      "description" -> s"$ditherNm.000 nm, $q.000000″".asJson,
-      "observeClass" -> "SCIENCE".asJson,
-      "steps" -> steps.asJson
-    )
 
   def addEndStepEvent(sid: Step.Id): IO[Unit] = {
     val q = s"""
